split error file
