+++ packages/client/src/prax.ts
@@ -12,6 +12,7 @@ import { createPromiseClient } from '@connectrpc/connect';
 import { createChannelTransport } from '@penumbra-zone/transport-dom/create';
 import { PenumbraSymbol } from '.';
 import { jsonOptions } from '@penumbra-zone/protobuf';
+import { PraxNotAvailableError, PraxNotConnectedError, PraxNotInstalledError } from './error';
 
 const prax_id = 'lkpmkhpnhknhmibgnmmhdhgdilepfghe';
 const prax_origin = `chrome-extension://${prax_id}`;
@@ -73,43 +74,3 @@ export const createPraxTransport = () => createChannelTransport(praxTransportOpt
 let praxTransport: Transport | undefined;
 export const createPraxClient = <T extends PenumbraService>(service: T): PromiseClient<T> =>
   createPromiseClient(service, (praxTransport ??= createPraxTransport()));
-
-export class PraxNotAvailableError extends Error {
-  constructor(
-    message = 'Prax not available',
-    public opts?: ErrorOptions,
-  ) {
-    super(message, opts);
-    this.name = 'PraxNotAvailableError';
-  }
-}
-
-export class PraxNotConnectedError extends Error {
-  constructor(
-    message = 'Prax not connected',
-    public opts?: ErrorOptions,
-  ) {
-    super(message, opts);
-    this.name = 'PraxNotConnectedError';
-  }
-}
-
-export class PraxNotInstalledError extends Error {
-  constructor(
-    message = 'Prax not installed',
-    public opts?: ErrorOptions,
-  ) {
-    super(message, opts);
-    this.name = 'PraxNotInstalledError';
-  }
-}
-
-export class PraxManifestError extends Error {
-  constructor(
-    message = 'Incorrect Prax manifest href',
-    public opts?: ErrorOptions,
-  ) {
-    super(message, opts);
-    this.name = 'PraxManifestError';
-  }
-}
